fix(reputation): update return type for feedback index and enhance contract interface
- Changed the return type of the `getFeedbackIndex` function in `IReputationRegistry` from `uint64` to `uint256` for improved compatibility.
- Updated `ReputationRegistryFacet` to implement `IReputationRegistry`, ensuring adherence to the interface.
- Added a new function `getIdentityRegistry` to return the address of the current contract.

These changes enhance the functionality and type safety of the reputation management system.

+++ packages/contracts/src/apps/facets/reputation/IReputationRegistry.sol
@@ -159,7 +159,7 @@ interface IReputationRegistry is IReputationRegistryBase {
         address clientAddress,
         uint64 feedbackIndex,
         address[] calldata responders
-    ) external view returns (uint64);
+    ) external view returns (uint256);
 
     /// @notice Gets the clients
     /// @param agentId The ID of the agent
